@@ -1,14 +1,15 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.1.8'
-gem 'pg'
-gem 'sass-rails', '~> 4.0.3'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.0.0'
+gem 'rails', '4.2.11'
+gem 'pg', "~> 0.18"
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
 gem 'jquery-rails'
 gem 'turbolinks'
-gem 'jbuilder', '~> 2.0'
-gem 'sdoc', '~> 0.4.0',          group: :doc
+gem 'jbuilder'
+gem 'sdoc',         group: :doc
+gem 'eventmachine'
 
 gem 'decent_exposure'
 
@@ -16,17 +17,17 @@ gem 'haml-rails'
 gem 'bootstrap-sass'
 gem 'simple_form'
 gem 'devise'
-gem 'sprig', '~> 0.1'
+gem 'sprig'
 
 group :development, :test do
   gem 'bogus', require: false
-  gem 'capybara'
+  gem 'capybara', '>= 2.0'
+  gem 'capybara-webkit', '>= 1.15'
   gem 'database_cleaner'
-  gem 'factory_girl_rails'
+  gem 'factory_bot'
   gem 'ffaker'
   gem 'poltergeist', require: false
   gem 'selenium-webdriver'
-  gem 'capybara-webkit'
   gem 'pry'
   gem 'pry-rails'
   gem 'quiet_assets'
